+++ src/Triangle.java
@@ -1,10 +1,10 @@
 public class Triangle {
-    float a;
-    float b;
-    float c;
+    double a;
+    double b;
+    double c;
 
-    public float area() {
-        float p = 0.5*(a+b+c);
+    public double area() {
+        double p = 0.5*(a+b+c);
         return Math.sqrt(p*(p-a)*(p-b)*(p-c));
     }
 }
