+++ src/circle.java
@@ -1,5 +1,5 @@
 public class circle {
     double R;
 
-    public float area() { return 3.14*R*R; }
+    public double area() { return 3.14*R*R; }
 }
